Fix GitHub rate limit and ETag handling
